feat(chat): add versions for applications to create publication request form (Issue #2025)
