+++ apps/chat/src/components/Common/ReplaceConfirmationModal/Components.tsx
@@ -447,21 +447,23 @@ export const FilesRow = ({
 
 interface ApplicationViewProps {
   item: ShareEntity;
-  onSelect?: (ids: string[]) => void;
   isChosen?: boolean;
+  featureContainerClassNames?: string;
+  onSelect?: (ids: string[]) => void;
 }
 
 export interface ApplicationRowProps extends ApplicationViewProps {
   level?: number;
-  onEvent?: (eventId: ReplaceOptions, data: string) => void;
   additionalItemData?: Record<string, unknown>;
   itemComponentClassNames?: string;
+  onEvent?: (eventId: ReplaceOptions, data: string) => void;
 }
 
 const ApplicationView = ({
   item: application,
-  onSelect,
   isChosen,
+  featureContainerClassNames,
+  onSelect,
 }: ApplicationViewProps) => {
   const entity = {
     ...application,
@@ -470,7 +472,7 @@ const ApplicationView = ({
   };
 
   return (
-    <FeatureContainer>
+    <FeatureContainer containerClassNames={featureContainerClassNames}>
       {onSelect && (
         <div
           className="relative flex size-[18px] shrink-0"
@@ -515,9 +517,10 @@ export const ApplicationRow = ({
   level,
   item: application,
   additionalItemData,
-  onEvent,
   itemComponentClassNames,
   isChosen,
+  featureContainerClassNames,
+  onEvent,
   onSelect,
 }: ApplicationRowProps) => {
   return (
@@ -531,6 +534,7 @@ export const ApplicationRow = ({
     >
       <ApplicationView
         isChosen={isChosen}
+        featureContainerClassNames={featureContainerClassNames}
         onSelect={onSelect}
         item={application}
       />
